Moved multisrc to src and stripped some code
