--- lib-multisrc/pawchive/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    alias(kei.plugins.multisrc)
-}
-
-keiyoushi {
-    baseVersionCode = 1
-    libVersion = "1.4"
-}
