+++ src/all/pawchive/build.gradle.kts
@@ -1,18 +1,17 @@
 import io.github.keiyoushi.gradle.api.ContentWarning
 
 plugins {
-	alias(kei.plugins.extension)
+    alias(kei.plugins.extension)
 }
 
 keiyoushi {
-	name = "Pawchive"
-	versionCode = 1
-	contentWarning = ContentWarning.NSFW
-	libVersion = "1.4"
-	theme = "pawchive"
+    name = "Pawchive"
+    versionCode = 1
+    contentWarning = ContentWarning.NSFW
+    libVersion = "1.4"
 
-	source {
-		lang = "all"
-		baseUrl = "https://pawchive.pw"
-	}
+    source {
+        lang = "all"
+        baseUrl = "https://pawchive.pw"
+    }
 }
