+++ src/all/pawchive/res/mipmap-mdpi/ic_launcher.png
